refactor(core): move preset middlewares to dedicated preset folder
Move all preset-related middlewares from middlewares/system to middlewares/preset for better organization and separation of concerns.
