+++ packages/core/src/middlewares/preset/delete_preset.ts
